repo: moved header files to include folder and restructure folders
