+++ include/algorithm/generic/common/types.hpp
