+++ src/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(generic)
+add_subdirectory(robotics)
